refactor:
- Add inline script metadata (PEP 723) to main.py and stats.py
-Removal of pyproject.toml and uv.lock
- Convert relative imports to absolute imports in 'script/' to support standalone execution
- Update TOML paths
- Add instructions in README.md for running statistics
