--- scripts/imports/fdc/pyproject.toml
@@ -1,12 +1,0 @@
-[project]
-name = "fdc-import"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.14"
-dependencies = []
-
-[dependency-groups]
-dev = [
-    "ijson>=3.5.0",
-]
